Added Score pop ups
Eaten ghost now give score with a streak of max 4 200->400->800->1600
each streak with a custom popup

Level now automatically change when all pellets are picked up

All levels are now stored in 1 Json. with level changing happening in the grid.

Ghosts and MsPacman gets reset when dying or progressing to a new level.